pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cp2077_voiceswap"
version = "0.0.1"
authors = [
  { name="Zhincore", email="adam@zhincore.eu" },
]
description = ""
readme = "README.md"
requires-python = ">=3.7"
dependencies  = [
  "asyncio",
  "python-dotenv"
]

[project.optional-dependencies]
dev = ["isort", "autopep8"]

[project.scripts]
voiceswap = "main:main"
extract_files = "main:extract_files"
wem2ogg = "main:wem2ogg"

[project.urls]
"Homepage" = "https://github.com/zhincore/cp2077-voiceswap"
"Bug Tracker" = "https://github.com/zhincore/cp2077-voiceswap/issues"
